--- src/app/pages/technical-settings/crypto/crypto-settings.component.html
@@ -1,12 +1,0 @@
-<div>
-    <form class="form" [formGroup]="formGroup" height="100%">
-        <div class="mat-toolbar-row left-actions">
-            <button mat-raised-button color="warn" (click)="save(formGroup.value)"
-                [disabled]="!formGroup.valid || !formGroup.dirty">
-                <mat-icon>save</mat-icon><span>{{'general.save' | translate}}</span>
-            </button>
-        </div>
-        <app-crypto-key *ngIf="cryptoSettings?.crypto"
-      [cryptoSettings]="cryptoSettings" [formGroup]="formGroup"></app-crypto-key>
-    </form>
-</div>
