+++ src/app/pages/technical-settings/technical-settings.component.html
@@ -1,16 +1,14 @@
 <div class="main-content">
-    <mat-tab-group (selectedIndexChange)="updateRoute($event)" [selectedIndex]="activeTabIndex" animationDuration="0ms"
-        disableRipple="true" class="h-100 mat-tab-info">
-
-        <mat-tab>
-            <ng-template mat-tab-label>
-                <mat-icon>enhanced_encryption</mat-icon><span>{{'technical_settings.tabs.crypto' | translate}}</span>
-            </ng-template>
-            <ng-template matTabContent>
-                <app-crypto-settings></app-crypto-settings>
-            </ng-template>
-        </mat-tab>
-
-    </mat-tab-group>
+  <mat-tab-group (selectedIndexChange)="updateRoute($event)" [selectedIndex]="activeTabIndex" animationDuration="0ms"
+      disableRipple="true" class="h-100 mat-tab-info">
+    <mat-tab>
+      <ng-template mat-tab-label>
+        <mat-icon>enhanced_encryption</mat-icon><span>{{'technical_settings.tabs.crypto' | translate}}</span>
+      </ng-template>
+      <ng-template matTabContent>
+        <app-settings-crypto></app-settings-crypto>
+      </ng-template>
+    </mat-tab>
+  </mat-tab-group>
 </div>
