+++ src/app/pages/technical-settings/technical-settings.routing.ts
@@ -5,12 +5,12 @@ import { Action, Entity } from '../../types/Authorization';
 import { TechnicalSettingsComponent } from './technical-settings.component';
 
 export const TechnicalSettingsRoutes: Routes = [
-    {
-        path: '', component: TechnicalSettingsComponent, canActivate: [RouteGuardService], data: {
-            auth: {
-                entity: Entity.SETTING,
-                action: Action.CREATE,
-            },
-        },
+  {
+    path: '', component: TechnicalSettingsComponent, canActivate: [RouteGuardService], data: {
+      auth: {
+        entity: Entity.SETTING,
+        action: Action.CREATE,
+      },
     },
+  },
 ];
